Fix smoketest terminating terminated instances

+++ upload-ami/src/upload_ami/smoke_test.py
@@ -55,7 +55,8 @@ def smoke_test(image_id, run_id, cancel):
         raise
     finally:
         logging.info(f"Terminating instance {instance_id}")
-        ec2.terminate_instances(InstanceIds=[instance_id])
+        if instance["State"]["Name"] != "terminated":
+            ec2.terminate_instances(InstanceIds=[instance_id])
         ec2.get_waiter("instance_terminated").wait(InstanceIds=[instance_id])
